chore: update custom element manifest scripts to simpify
