+++ package.json
@@ -12,20 +12,18 @@
         "analyze": "lit-analyzer \"{packages,tools}/*/src/**/!(*.css).ts\"",
         "build": "wireit",
         "build:clear-cache": "rimraf packages/*/tsconfig.tsbuildinfo && rimraf tools/*/tsconfig.tsbuildinfo",
-        "build:confirm": "node ./scripts/confirm-build.js",
         "build:css": "wireit",
         "build:css:watch": "wireit",
-        "build:react": "yarn gen-react-wrapper && node ./scripts/build-react.js && yarn tsc --build tsconfig-react-wrapper.json",
+        "build:react": "rimraf react && node ./scripts/build-react.js && tsc --build tsconfig-react-wrapper.json",
         "build:tests": "tsc --build test/tsconfig.json && tsc --build test/tsconfig-node.json",
         "build:ts": "wireit",
         "build:ts:watch": "wireit",
         "build:types": "wireit",
         "build:watch": "wireit",
-        "changeset-snapshot-publish": "yarn prepublishOnly && yarn changeset version --snapshot snapshot && yarn constraints --fix && yarn update-version && yarn changeset publish --no-git-tag --tag snapshot",
-        "changeset-publish": "yarn prepublishOnly && yarn changeset version && yarn install && yarn constraints --fix && yarn update-version && yarn changeset publish --no-git-tag && yarn push-to-remote && yarn create-git-tag && yarn postpublish",
-        "update-version": "node ./scripts/update-version.js",
+        "publish:changeset-snapshot": "yarn prepublishOnly && yarn changeset version --snapshot snapshot && yarn update-version && yarn changeset publish --no-git-tag --tag snapshot",
+        "publish:changeset": "yarn prepublishOnly && yarn changeset version && yarn update-version && yarn changeset publish --no-git-tag && yarn push-to-remote && node --no-warnings ./scripts/create-git-tag.js && yarn postpublish",
+        "update-version": "node ./scripts/update-version.js && yarn constraints --fix && yarn install --refresh-lockfile",
         "chromatic": "chromatic --build-script-name storybook:build # note that --project-token must be set in your env variables",
-        "create-git-tag": "node --no-warnings scripts/create-git-tag.js",
         "custom-element-json": "node ./scripts/custom-element-json.js",
         "docs:analyze": "cem analyze --globs \"packages/**/*.ts\" --exclude \"**/*.d.ts\" --exclude \"**/stories/**\" --exclude \"**/icons/**\" --exclude \"**/elements/**\" --outdir projects/documentation --litelement",
         "docs:build": "yarn workspace documentation build",
@@ -35,21 +33,18 @@
         "docs:review": "alex packages/**/*.md",
         "docs:start": "yarn workspace documentation serve --watch",
         "find": "test -f custom-elements.json",
-        "gen-react-wrapper": "node ./scripts/gen-react-wrapper.js",
         "icons": "wireit",
         "icons:ui": "wireit",
         "icons:workflow": "wireit",
         "lint": "git status --porcelain && git add . && lint-staged --allow-empty",
         "new-package": "cd projects/templates && plop",
-        "postcustom-element-json": "node ./scripts/run-check-cem.js",
         "postinstall": "husky || true && patch-package",
         "postpublish": "yarn prepublish:react && yarn publish:react && yarn postpublish:react",
         "postpublish:react": "git reset --hard HEAD^ && git prune && rimraf react",
-        "precustom-element-json": "node ./scripts/remove-custom-elements-json.js",
         "preeleventy": "yarn docs:analyze",
         "prepack": "pinst --disable",
         "postpack": "pinst --enable",
-        "prepublishOnly": "rimraf react && yarn build && yarn custom-element-json && yarn build:confirm",
+        "prepublishOnly": "yarn build && yarn custom-element-json && node ./scripts/confirm-build.js",
         "prepublish:react": "yarn build:react && sed -i \"\" \"s/react/# react/g\" .gitignore && git commit -am \"Commit React Wrappers\" --no-verify",
         "prestorybook": "wireit",
         "prestorybook:build": "cem analyze --outdir storybook/",
@@ -177,7 +172,7 @@
         "prettier-plugin-package": "^1.3.0",
         "pretty-bytes": "^7.0.0",
         "re-template-tag": "^2.0.1",
-        "rimraf": "^5.0.1",
+        "rimraf": "^6.0.1",
         "rollup": "^4.12.0",
         "sinon": "^17.0.1",
         "storybook": "^8.6.12",
