+++ projects/documentation/package.json
@@ -49,7 +49,7 @@
         "posthtml-attrs-parser": "^1.0.1",
         "posthtml-match-helper": "^2.0.0",
         "prismjs": "^1.29.0",
-        "rimraf": "^5.0.1",
+        "rimraf": "^6.0.1",
         "rollup": "^4.12.0",
         "rollup-plugin-lit-css": "^5.0.0",
         "rollup-plugin-minify-html-literals": "^1.2.6",
