+++ projects/example-project-rollup/package.json
@@ -39,7 +39,7 @@
     "eslint": "^8.23.0",
     "eslint-config-prettier": "^9.1.0",
     "prettier": "^3.5.3",
-    "rimraf": "^5.0.1",
+    "rimraf": "^6.0.1",
     "rollup": "^4.12.0",
     "rollup-plugin-styles": "^4.0.0",
     "rollup-plugin-visualizer": "^5.12.0",
